# docs/swagger.yaml
basePath: ./
definitions:
...
  dto.Error:
    properties:
      message:
        type: string
    type: object
  dto.IDWrapper:
    properties:
      id:
        minimum: 1
        type: integer
    required:
    - id
    type: object
...
  dto.PaginatedProduct:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
      total:
        description: This will always return the total of all records
        type: integer
    type: object
...
  dto.Product:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      discount_price:
        type: number
      id:
        type: integer
      name:
        type: string
      specifications:
        type: string
      status_id:
        type: integer
      supplier_id:
        type: integer
      tags:
        type: string
      unit_price:
        type: number
    type: object
...
  dto.ProductUpdate:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      discount_price:
        type: number
      id:
        type: integer
      name:
        type: string
      specifications:
        type: string
      status_id:
        type: integer
      supplier_id:
        type: integer
      tags:
        type: string
      unit_price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Describes technoStore REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: technoStore API
  version: "1.0"
paths:
...
  /v1/product:
    post:
      consumes:
      - application/json
      description: Create a new product in the system
      parameters:
      - description: Product params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IDWrapper'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Add a new product
      tags:
      - Product
...
  /v1/product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Product by id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Product delete processed
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Delete a Product by id
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get a Product by id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Product'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Get a Product by id
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Update a product by id
      parameters:
      - description: product params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProductUpdate'
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: productDto updated
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Update a product by id
      tags:
      - Product
  /v1/products:
    get:
      consumes:
      - application/json
      description: Get Products by query
      parameters:
      - description: q is query string of product name
        in: query
        name: q
        type: string
      - collectionFormat: multi
        description: brand
        in: query
        items:
          type: integer
        name: brand
        type: array
      - description: category
        in: query
        name: category
        type: integer
      - description: supplier
        in: query
        name: supplier
        type: integer
      - description: verified_supplier
        in: query
        name: verified_supplier
        type: boolean
      - description: min_price
        in: query
        name: min_price
        type: number
      - description: max_price
        in: query
        name: max_price
        type: number
      - description: sort
        in: query
        name: sort
        type: string
      - description: order
        in: query
        name: order
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedProduct'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Get Products by query
      tags:
      - Product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"